[HLSL] Move the D3D12 linear algebra shim into LinAlgAbi.h
The local stand-ins for the linear algebra API were spread across three places:
the enumerations in HlslExecTestUtils.h, the capability structures in
HlslExecTestUtils.cpp, and the two D3D12_FEATURE identifiers among the
implementation constants. Nothing marked the boundary, so it read as arbitrary,
and it left the same explanatory comment in two places and two TODOs to action
when the types ship.

It was also load-bearing in a way nobody would guess. Because the namespace was
opened twice, the block in the .cpp had to stay at global scope: nesting it in
the anonymous namespace below it would have declared a second, distinct
linalg_abi that silently shadowed the header's. That hazard existed only
because the shim was split.

It is now one file, LinAlgAbi.h, included by HlslExecTestUtils.h. The
enumerations have to be visible there regardless - they are the field types of
the capability queries - so this also keeps ninety-odd lines of linear algebra
mirrors out of a header that ExecutionTest.cpp and LongVectors.cpp include
without needing any of it.

The ASSERT_RUNTIME_* blocks move with it, so the file carries its own proof
rather than being verified from somewhere else, and deleting it once the API
ships leaves nothing dangling.

Feature identifiers

D3D12_FEATURE_LINEAR_ALGEBRA_SUPPORT and
D3D12_FEATURE_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT were spelled
LinearAlgebraSupportFeature and LinearAlgebraOperationSupportFeature, which is
the one part of the shim that did not say what it mirrored. They now use their
D3D12 names like everything else. D3D12_FEATURE itself ships in the released
SDK, so they stay constants rather than becoming an enumeration.

That also folds their two hand-written assertions into the ASSERT_RUNTIME_ENUM
list, and removes a stale undef of ASSERT_RUNTIME_OFFSET_SAME, which no longer
exists.

No functional change: the types, values and layouts are the same, and no
CMakeLists change is needed as only .cpp files are listed.

Validation

Build, clang-format 17.0.6 and git diff --check are clean, and abicheck exits 0.

The assertions were confirmed still armed after the move rather than assumed,
with three separate negative controls: setting the local FLOAT16 to 6, setting
the feature identifier to 76, and swapping two struct fields. Each fails the
build at the expected line in LinAlgAbi.h, so the value, the newly mirrored
feature identifier and the field offsets all still reach the real header.

TAEF on WARP is Total=24, Passed=19, Failed=4, Skipped=1, with a non-passing set
identical to the previous commit. The four failures and one skip are the known
runtime spec-lag cases and are unrelated to this change.

Copilot-Session: 83725f5d-8e98-4c1d-91ee-ad47629e007b
